+++ backend/src/main/java/com/team5/secondhand/BackendApplication.java
@@ -1,4 +1,4 @@
-package com.team5.backend;
+package com.team5.secondhand;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
