+++ backend/src/test/java/com/team5/secondhand/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.team5.backend;
+package com.team5.secondhand;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
